+++ packages/xo-web-6/src/App/index.tsx
@@ -7,7 +7,10 @@ import { withStore } from 'reaclette'
 const Bar = React.lazy(() => import('./Bar'))
 const Foo = React.lazy(() => import('./Foo'))
 const Visualization = React.lazy(() => import('./Visualization'))
+const VisualizationHost = React.lazy(() => import('./Visualization/hostStats'))
 const Visualization2 = React.lazy(() => import('./Visualization/mini-stats'))
+const VisualizationSr = React.lazy(() => import('./Visualization/storageStats'))
+
 //import Visualization2 from './Visualization/mini-stats'
 
 
@@ -61,8 +64,10 @@ export default withStore(
         <React.Suspense fallback='loading'>
           <Route path='/' exact component={Bar} />
           <Route path='/foo' component={Foo} />
-          <Route path='/visualization' component={Visualization} />
-          <Route path='/visualization2' component={Visualization2} />
+          <Route path='/vm' component={Visualization} />
+          <Route path='/miniStats' component={Visualization2} />
+          <Route path='/host' component={VisualizationHost}/>
+          <Route path='/storage' component={VisualizationSr}/>
         </React.Suspense>
       </HelmetProvider>
     </Router>
